processing/requirements.txt
appnope==0.1.0
argon2-cffi==20.1.0
async-generator==1.10
#attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1605083924122/work
backcall==0.2.0
bleach==3.2.1
#blis @ file:///opt/concourse/worker/volumes/live/f9b87811-6932-46c2-6826-5aef90cfc1ba/volume/cython-blis_1594246874825/work
brotlipy==0.7.0
#catalogue @ file:///Users/runner/miniforge3/conda-bld/catalogue_1605613876625/work
certifi==2020.11.8
#cffi @ file:///opt/concourse/worker/volumes/live/8fbed55e-0573-4246-7cf9-162067141d5e/volume/cffi_1605538070636/work
#chardet @ file:///Users/runner/miniforge3/conda-bld/chardet_1602255309528/work
click==7.1.2
#cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1604179103599/work
#cymem @ file:///opt/concourse/worker/volumes/live/4097c5b0-7ead-4af2-417f-f93ff176ec70/volume/cymem_1605062271900/work
decorator==4.4.2
defusedxml==0.6.0
elasticsearch==7.10.0
entrypoints==0.3
#idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
#importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1606196818768/work
influxdb==5.2.3
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kafka-python==1.4.7
lxml==4.6.1
MarkupSafe==1.1.1
mistune==0.8.4
murmurhash==1.0.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nl-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/nl_core_news_lg-2.3.0/nl_core_news_lg-2.3.0.tar.gz
#nl-core-news-md @ https://github.com/explosion/spacy-models/releases/download/nl_core_news_md-2.3.0/nl_core_news_md-2.3.0.tar.gz
#nl-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/nl_core_news_sm-2.3.0/nl_core_news_sm-2.3.0.tar.gz
#nltk==3.5
#notebook==6.1.5
numpy
oauthlib==3.1.0
packaging==20.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
plac==1.1.3
preshed==3.0.4
prometheus-client==0.9.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
#pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
Pygments==2.7.2
pymongo==3.11.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
#pyrsistent @ file:///Users/runner/miniforge3/conda-bld/pyrsistent_1605115614071/work
#PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1602326916036/work
python-dateutil==2.8.1
pytz==2020.4
pyzmq==20.0.0
qtconsole==5.0.0
QtPy==1.9.0
regex==2020.11.13
requests
requests-oauthlib==1.3.0
Send2Trash==1.5.0
#six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
spacy
spacy-lookups-data==0.3.2
srsly==1.0.4
terminado==0.9.1
testpath==0.4.4
#thinc @ file:///opt/concourse/worker/volumes/live/d0d401d5-f84a-4764-440e-d67893479983/volume/thinc_1594251950205/work
tornado==6.1
#tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1606103150798/work
traitlets==5.0.5
tweepy==3.9.0
#urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1603125704209/work
#wasabi @ file:///home/conda/feedstock_root/build_artifacts/wasabi_1600272362626/work
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
#zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1603668650351/work
neo4j
pyyaml
